+++ tests/FreeElephants/JsonApi/DTO/ResourceObjectTest.php
@@ -30,6 +30,29 @@ class ResourceObjectTest extends AbstractTestCase
         $this->assertInstanceOf(Example\OneRelationships::class, $resourceObject->relationships);
         $this->assertSame('one', $resourceObject->relationships->one->data->type);
     }
+
+    public function testWithNullableId(): void
+    {
+        $resourceObject = new class([
+            'type'          => 'type',
+            'attributes'    => [
+                'foo' => 'bar',
+            ],
+            'relationships' => [
+                'one' => [
+                    'data' => [
+                        'type' => 'one',
+                        'id'   => 'one',
+                    ],
+                ],
+            ],
+        ]) extends AbstractResourceObject {
+            public Example\Attributes $attributes;
+            public Example\OneRelationships $relationships;
+        };
+
+        $this->assertNull($resourceObject->id);
+    }
 }
 
 class Attributes extends AbstractAttributes